+++ src/App.tsx
@@ -1,26 +1,6 @@
-import React from "react";
-import { RouterProvider, createBrowserRouter } from "react-router-dom";
-import Layout from "./components/layout/Layout";
+import { RouterProvider } from "react-router-dom";
 import { GlobalStyle } from "./style/global";
-import Home from "./pages/Home";
-import { Error } from "./components/common/Error";
-
-const routerList = [
-  {
-    path: "/",
-    element: <Home />,
-  },
-];
-
-const router = createBrowserRouter(
-  routerList.map((item) => {
-    return {
-      ...item,
-      element: <Layout>{item.element}</Layout>,
-      errorElement: <Error />,
-    };
-  })
-);
+import { router } from "./router";
 
 function App() {
   return (
